requirements.txt
tweepy-3.8.0
PyQt5
schedule